% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.R
\name{AF}
\alias{AF}
\title{Adaptive Fisher Test}
\usage{
AF(
  Y,
  X,
  binary = FALSE,
  cov = NULL,
  nperm = 1000,
  adapt_perm = FALSE,
  cutoff = 2.5e-06,
  n0 = 1,
  seed = NULL
)
}
\arguments{
\item{Y}{Response Variable. A vector or length \code{n} (the number of samples).}

\item{X}{Explanatory Variables to be tested. A matrix with dimensions \code{n} by \code{K}
(the number of explanatory variables to be tested).}

\item{binary}{Indicator of whether \code{Y} is binary.}

\item{cov}{Covariates. A matrix with dimensions \code{n}
by \code{J} (the number of explanatory variables that are not to be tested).}

\item{nperm}{Number of permutations. Also the starting number of
permutations for "step-up" algorithm. Default is \code{1,000}.}

\item{adapt_perm}{Whether "step-up" algorithm is used for P-value
calculation. If FALSE, function permutes \code{nperm} times and stops.
If TRUE, \code{nperm} will be increased \code{10} times each round if P-value
<= \code{5/nperm}. Algorithm stops if P-value > \code{5/nperm} or <= \code{cutoff}.}

\item{cutoff}{Cutoff for "step-up" algorithm.}

\item{n0}{Tuning parameter. Discard the first \code{n0-1} P-values of each sample.}

\item{seed}{Specify the seed for permutations.}
}
\value{
An object of "AF" class.
\describe{
 \item{pv}{P-value of AF test.}
 \item{stat}{AF statisitc.}
 \item{indexes}{Indexes of P-values combined into the test
  statistic. Indexes are sorted so that P-values are
  in ascending order.}
  \item{stat_all}{AF statistics for all permuted samples.}
  \item{pv_all}{P-values of AF statistics for all permuted samples.}
  \item{method}{Method used.}
}
}
\description{
This function performs Adaptive Fisher (AF) test on GLM-based
Score Statistics.
}
\examples{
Y <- bs_dense$trait
methyl <- bs_dense$methyl
pos <- bs_dense$pos
test <- AF(Y, methyl, binary = TRUE)
summary(test)

}
\seealso{
\code{\link{set.seed}}
}
