% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_combine.R
\name{AF_combine}
\alias{AF_combine}
\title{AF Combination}
\usage{
AF_combine(p, log = FALSE, weight = NULL, n0 = 1)
}
\arguments{
\item{p}{P-values to be combined. A matrix with dimensions \code{K} by \code{N}.
If an object of P-values from perm_score is used,
\code{N} is the number of permutations plus 1.}

\item{log}{Indicator of whether P-values are on the log scale.}

\item{weight}{Weights given to the P-values. A vector with dimension
\code{K}. Constant weights are used if it is not specified.}

\item{n0}{Tuning parameter. Discard the first \code{n0-1} P-values of each
sample.}
}
\value{
A list object.
\describe{
  \item{pv}{P-value of AF test.}
  \item{stat}{AF statistic.}
  \item{indexes}{Indexes of P-values combined into the test
  statistic. Indexes are sorted so that P-values are
  in ascending order.}
  \item{stat_all}{AF statistics for all permuted samples.}
  \item{pv_all}{P-values of AF statistics for all permuted samples.}
}
}
\description{
This function combines P-values using
Adaptive Fisher (AF) method.
}
\examples{
# Combine P-values of normally distributed test statistics
U <- matrix(rnorm(10 * 100), ncol=100)
p <- 2 * (1 - pnorm(abs(U)))
wt <- (1:10)/55
test <- AF_combine(p, weight = wt)

}
