% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFpca.R
\name{AFpca}
\alias{AFpca}
\title{Adaptive Fisher Method with Principal Components (AFpca)
for Trait-Methylation Set Association}
\usage{
AFpca(
  Y,
  M,
  binary = FALSE,
  cov = NULL,
  varprop = 0.95,
  nperm = 1000,
  n0 = 1,
  adapt_perm = FALSE,
  cutoff = 2.5e-06,
  seed = NULL
)
}
\arguments{
\item{Y}{Y Phenotype data. It can be a continuous trait or a binary trait.
A vector of length \code{n} (number of subjects).}

\item{M}{A matrix of methylation levels with dimensions \code{n} by \code{K}
(\code{n} subjects, \code{K} CpG sites).}

\item{binary}{Indicator of whether \code{Y} is binary.}

\item{cov}{Covariates. A matrix with dimensions \code{n}
by \code{J} (number of covariates).}

\item{varprop}{Cutoff for proportion of variance to be expalined.
The first Kp principal components (PCs) are chosen such that \code{varprop}
of the total variance is explained.}

\item{nperm}{Number of permutations. Also the starting number of permutations
for "step-up" algorithm. Default is \code{1,000}.}

\item{n0}{Tuning parameter. Discard the first \code{n0-1} P-values of each sample.}

\item{adapt_perm}{Whether "step-up" algorithm is used for P-value
calculation. If FALSE, function permutes \code{nperm} times and stops. If TRUE,
\code{nperm} will be increased \code{10} times each round if P-value <= \code{5/nperm}.
Algorithm stops if P-value > \code{5/nperm} or <= \code{cutoff}.}

\item{cutoff}{Cutoff for "step-up" algorithm.}

\item{seed}{Specify the seed for permutations.}
}
\value{
An object of "AFpca" class.
\describe{
 \item{pv}{P-value of AFb test.}
 \item{stat}{Test statistic of AFb test.}
 \item{indexes}{Indexes of PCs combined into the test
  statistic. Indexes are sorted so that P-values are
  in ascending order.}
  \item{stat_all}{AFb statistics for all permuted samples.}
  \item{pv_all}{P-values of AFb statistics for all permuted samples.}
  \item{method}{Method used.}
}
}
\description{
Adaptive Fisher Method with Principal Components (AFpca)
for Trait-Methylation Set Association
}
\examples{
Y <- bs_dense$trait
methyl <- bs_dense$methyl
pos <- bs_dense$pos
test <- AFpca(Y, methyl, varprop = 0.9,
            binary = TRUE, adapt_perm = TRUE)
summary(test)

}
