% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFv.R
\name{AFv}
\alias{AFv}
\title{Adaptive Fisher Method for Testing Methylation Variation}
\usage{
AFv(
  Y,
  M,
  pos,
  nbasis,
  start = NULL,
  end = NULL,
  binary = FALSE,
  cov = NULL,
  nperm = 1000,
  bandwidth = 5509,
  kernel = c("normal", "box"),
  adapt_perm = FALSE,
  cutoff = 2.5e-06,
  seed = NULL,
  n0 = 1,
  ...
)
}
\arguments{
\item{Y}{Phenotype data. It can be a continuous trait or a binary trait.
A vector of length \code{n} (number of subjects).}

\item{M}{A matrix of methylation levels with dimensions \code{n} by \code{K}
(\code{n} subjects, \code{K} CpG sites).}

\item{pos}{A vector of CpG locations (in base pairs). Elements
should be of the same order as the columns of \code{M}.}

\item{nbasis}{The number of B-spline basis functions.}

\item{start}{Start location of the region.}

\item{end}{End location of the region.}

\item{binary}{Indicator of whether \code{Y} is binary.}

\item{cov}{Covariates. A matrix with dimensions \code{n}
by \code{J} (number of covariates).}

\item{nperm}{Number of permutations. Also the starting number of permutations
for "step-up" algorithm. Default is \code{1,000}.}

\item{bandwidth}{bandwidth The bandwidth for \code{ksmooth}. Default is \code{5509},
20 \% of the median gene length according to UCSC Genome build hg38.}

\item{kernel}{kernel The kernel to be used for \code{ksmooth}}

\item{adapt_perm}{Whether "step-up" algorithm is used for P-value
calculation. If FALSE, function permutes \code{nperm} times and stops. If TRUE,
\code{nperm} will be increased 10 times each round if P-value <= \code{5/nperm}.
Algorithm stops if P-value > \code{5/nperm} or <= \code{cutoff}.}

\item{cutoff}{Cutoff for "step-up" algorithm.}

\item{seed}{Specify the seed for permutations.}

\item{n0}{Tuning parameter. Discard the first \code{n0-1} P-values of each sample.}

\item{...}{Optional arguments \code{create.bspline.basis}.}
}
\value{
An object of "AFv" class.
\describe{
 \item{pv}{P-value of AFv test.}
 \item{stat}{Test statistic of AFv test.}
 \item{indexes}{Indexes of basis functions combined into the test
  statistic. Indexes are sorted so that P-values are
  in ascending order.}
  \item{stat_all}{AFv statistics for all permuted samples.}
  \item{pv_all}{P-values of AFv statistics for all permuted samples.}
  \item{method}{Method used.}
}
}
\description{
Adaptive Fisher Method for Testing Methylation Variation
}
\examples{
Y <- bs_dense$trait
methyl <- bs_dense$methyl
pos <- bs_dense$pos
test <- AFv(Y, methyl, pos, nbasis =10, bandwidth = 500,
            kernel = "box", binary = TRUE)
summary(test)

}
\seealso{
\code{\link[fda]{create.bspline.basis}},
\code{\link{set.seed}}, \code{\link[stats]{ksmooth}}
}
