% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{ar_dense}
\alias{ar_dense}
\title{Simulated Methylation Data with Autoregressive Effects in Dense Scenario}
\format{
A list object.
\describe{
  \item{trait}{A vector of length 500 containing disease labels
  (251 cases and 249 controls).}
  \item{methyl}{A 500 by 81 matrix containing methylation levels. Each
  row represents the methylation profile for one subject. Each
  element of the matrix denotes BS-seq methylation level at a CpG site.}
  \item{pos}{Locations for CpG sites (in base pairs).}
}
}
\usage{
data(ar_dense)
}
\description{
A simulated dataset containing 500 subjects and 81 methylation sites.
Methylation effects are constructed by first-order autoregressive model.
}
\keyword{datasets}
