% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisMat.R
\name{basisMat}
\alias{basisMat}
\title{Methylation Levels to Basis Function Values}
\usage{
basisMat(M, pos, nbasis, start = NULL, end = NULL, ...)
}
\arguments{
\item{M}{A matrix of methylation levels with dimensions \code{n} by \code{K}
(\code{n} subjects, \code{K} CpG sites).}

\item{pos}{A vector of CpG locations (in base pairs). Elements
should be of the same order as the columns of \code{M}.}

\item{nbasis}{The number of B-spline basis functions (default is cubic
splines). Note that \code{nbasis} should not be smaller than norder (default is 4).}

\item{start}{Start location of the region.}

\item{end}{End location of the region.}

\item{...}{Optional arguments for \code{create.bspline.basis}.}
}
\value{
A matrix of basis function values with dimensions \code{n} by \code{Kb},
where \code{Kb} is the number of basis functions.
}
\description{
Methylation Levels to Basis Function Values
}
\examples{
Y <- bs_dense$trait
methyl <- bs_dense$methyl
pos <- bs_dense$pos
X <- basisMat(methyl, pos, nbasis = 10)

}
\seealso{
\code{\link[fda]{create.bspline.basis}}
}
