% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispMat.R
\name{dispMat}
\alias{dispMat}
\title{Methylation Levels to Dispersion}
\usage{
dispMat(M, pos, bandwidth = 5509, kernel = c("normal", "box"))
}
\arguments{
\item{M}{A matrix of methylation levels with dimensions \code{n} by \code{K} (\code{n} subjects, \code{K} CpG sites).}

\item{pos}{A vector of CpG locations (in base pairs). Elements should be of the same
order as the columns of \code{M}.}

\item{bandwidth}{The bandwidth for \code{ksmooth}. Default is \code{5509},
20 \% of the median gene length according to UCSC Genome build hg38.}

\item{kernel}{The kernel to be used for \code{ksmooth}.}
}
\value{
A matrix of methylation dispersion with dimensions \code{n} by \code{K}.
}
\description{
Methylation Levels to Dispersion
}
\seealso{
\code{\link[stats]{ksmooth}}
}
