% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaMat.R
\name{pcaMat}
\alias{pcaMat}
\title{Methylation Levels to Principal Components}
\usage{
pcaMat(M, varprop = 0.95)
}
\arguments{
\item{M}{A matrix of methylation levels with dimensions \code{n} by \code{K}
(\code{n} subjects, \code{K} CpG sites).}

\item{varprop}{Cutoff for proportion of variance to be expalined.
The first \code{Kp} principal components (PCs) are chosen such that \code{varprop}
of the total variance is explained.}
}
\value{
A matrix of PCs with dimension \code{n} by \code{Kp}.
}
\description{
Methylation Levels to Principal Components
}
