% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_score.R
\name{perm_score}
\alias{perm_score}
\title{Permutation of GLM-based Score Statistics}
\usage{
perm_score(Y, X, binary = FALSE, cov = NULL, nperm = 1000, seed = NULL)
}
\arguments{
\item{Y}{Response Variable. A vector or length \code{n} (the number of samples).}

\item{X}{Explanatory Variables to be tested. A matrix with dimensions \code{n} by \code{K}
(the number of explanatory variables to be tested).}

\item{binary}{Indicator of whether \code{Y} is binary.}

\item{cov}{Covariates. A matrix with dimensions \code{n} (number of subjects)
by \code{J} (the number of explanatory variables that are not to be tested).}

\item{nperm}{Number of permutations. Default is \code{1,000}.}

\item{seed}{Specify the seed for permutations.}
}
\value{
A list object.
\describe{
  \item{Us}{Score statistics. The first column contains the original
  data.}
  \item{pvs}{P-values of two-tailed score tests;
  can be further used by AF_combine function.}
  \item{pvs_left}{Left-side P-values of one-tailed score tests;
  can be further used by AF_combine function.}
}
}
\description{
This function calculates and permutes GLM-based
score statistics.
}
\examples{
Y <- ar_dense$trait
X <- ar_dense$methyl
result <- perm_score(Y, X, binary = TRUE, nperm = 100)
names(result)

}
\seealso{
{\code{\link{set.seed}}}
}
