% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.AF}
\alias{print.AF}
\title{Print Function for Objects of "AF" Class}
\usage{
\method{print}{AF}(x, ...)
}
\arguments{
\item{x}{An object of "AF" class.}

\item{...}{Optional arguments for \code{print}.}
}
\value{
Method used; P-value; basis functions used;
CpG sites combined into the test statistic.
}
\description{
Print Function for Objects of "AF" Class
}
