% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.AFb}
\alias{print.AFb}
\title{Print Function for Objects of "AFb" Class}
\usage{
\method{print}{AFb}(x, ...)
}
\arguments{
\item{x}{An object of "AFb" class.}
}
\value{
Method used; P-value; basis functions used;
CpG sites combined into the test statistic.
}
\description{
Print Function for Objects of "AFb" Class
}
\examples{

}
