% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.AF}
\alias{summary.AF}
\title{Summary Function for Objects of "AF" Class}
\usage{
\method{summary}{AF}(object, ...)
}
\arguments{
\item{object}{An object of "AF" class.}

\item{...}{Optional arguments for \code{summary}.}
}
\value{
Method used; P-value; basis functions used;
CpG sites combined into the test statistic.
}
\description{
Summary Function for Objects of "AF" Class
}
