% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.AFb}
\alias{summary.AFb}
\title{Summary Function for Objects of "AFb" Class}
\usage{
\method{summary}{AFb}(object, ...)
}
\arguments{
\item{object}{An object of "AFb" class.}

\item{...}{Optional arguments for \code{summary}.}
}
\value{
Method used; P-value; basis functions used;
CpG sites combined into the test statistic.
}
\description{
Summary Function for Objects of "AFb" Class
}
