% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFb.R
\name{AFb}
\alias{AFb}
\title{Smoothed Adaptive Fisher (AFb) Test for Trait-Methylation Set Association}
\usage{
AFb(
  Y,
  M,
  pos,
  start,
  end,
  nbasis,
  binary = FALSE,
  cov = NULL,
  adapt_perm = FALSE,
  cutoff = 2.5e-06,
  nperm = 1000,
  seed = NULL,
  n0 = 1,
  ...
)
}
\arguments{
\item{Y}{Y Phenotype data. It can be a continuous trait or a binary trait.
A vector or length n (number of subjects).}

\item{M}{A matrix of methylation levels. A matrix with dimensions n by K
(n subjects, K methylation sites). Each row for a subject, and each column
for a methylated CpG site.}

\item{pos}{A vector of methylated site locations (in base pairs). Elements
should be of the same order as the columns of M.}

\item{start}{Start location of the region.}

\item{end}{End location of the region.}

\item{nbasis}{The number of B-spline basis functions.}

\item{binary}{Indicator of whether Y is binary.}

\item{cov}{Covariates. A matrix with dimensions n (number of subjects)
by J (number of covariates).}

\item{adapt_perm}{Whether "step-up" algorithm is used for P-value
calculation. If FALSE, function permutes nperm times and stops. If TRUE,
nperm will be increased 10 times each round if P-value <= 5/nperm.
Algorithm stops if P-value > 5/nperm or <= cutoff.}

\item{cutoff}{Cutoff for "step-up" algorithm.}

\item{nperm}{Number of permutations. Also the starting number of permutations
for "step-up" algorithm. Default is 1,000.}

\item{seed}{Specify random seed for permutations.}

\item{n0}{Tuning parameter. Discard the first n0-1 P-values of each column.}

\item{...}{Optional arguments \code{create.bspline.basis}.}
}
\value{
An object of "AFb" class.
\describe{
 \item{pv}{P-value of AFb test.}
 \item{stat}{Test statistic of AFb test.}
 \item{loci_combined}{Sites which are combined into the test
  statistic. The index of included sites are returned in the ascending
  order of their P-values.}
  \item{method}{Method used.}
}
}
\description{
Smoothed Adaptive Fisher (AFb) Test for Trait-Methylation Set Association
}
\examples{
Y <- bs_dense$trait
methyl <- bs_dense$methyl
pos <- bs_dense$pos
test <- AFb(Y, methyl, pos, start = min(pos), end = max(pos),
            nbasis = 50, binary = TRUE, adapt_perm = TRUE)
summary(test)

}
\seealso{
\code{\link[splines]{bs}}, \code{\link[fda]{fourier}},
\code{\link{set.seed}}
}
