% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{ar_dense}
\alias{ar_dense}
\title{Simulated Methylation Data with Autoregressive Effects in Dense Scenario}
\format{
A list object.
\describe{
  \item{trait}{A vector of length 1,000 containing disease labels
  (510 cases and 490 controls).}
  \item{methyl}{A 1,000 by 477 matrix containing methylation levels. Each
  row represents the methylaiton profile for one subject. Each
  element of the matrix denotes BS-seq methylation level at a CpG site.}
  \item{pos}{Locations for CpG sites (in base pairs).}
}
}
\usage{
data(ar_dense)
}
\description{
A simulated dataset containing 1,000 subjects and 477 methylation sites.
Methylation effects are constructed by first-order autoregressive model.
}
\keyword{datasets}
