% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisMat.R
\name{basisMat}
\alias{basisMat}
\title{Methylation Levels to Basis Function Values}
\usage{
basisMat(M, pos, basis = c("bs", "fourier"), ...)
}
\arguments{
\item{M}{A matrix of methylation levels. A matrix with dimensions n by K
(n subjects, K methylation sites). Each row for a subject, and each column
for a methylated CpG site.}

\item{pos}{A vector of methylated site locations (in base pairs). Elements
should be of the same order as the columns of M.}

\item{basis}{Basis functions used, "bs" for B-spline, "fourier" for Fourier.}

\item{...}{Optional arguments for constructing basis functions.}
}
\value{
A matrix of basis function values. A matrix with dimensions n by Kb,
where Kb is the number of basis functions.
}
\description{
Methylation Levels to Basis Function Values
}
\examples{
M <- ar_dense$methyl
pos <- ar_dense$pos
l <- length(pos)

## B-spline basis functions
knots <-seq(1, pos[l], length.out = 50)[-c(1, 50)]
X1 <- basisMat(M, pos, basis = "bs", degree = 2, knots = knots)

## Fourier basis functions
# if nbasis is an even number, function fourier takes nbasis = nbasis + 1
X2 <- basisMat(M, pos, basis = "fourier", nbasis = 20)

}
\seealso{
\code{\link[splines]{bs}}, \code{\link[fda]{fourier}}
}
