% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisMat.R
\name{basisMat}
\alias{basisMat}
\title{Methylation Levels to Basis Function Values}
\usage{
basisMat(M, pos, nbasis, start = NULL, end = NULL, ...)
}
\arguments{
\item{M}{A matrix of methylation levels. A matrix with dimensions n by K
(n subjects, K methylation sites). Each row for a subject, and each column
for a methylated CpG site.}

\item{pos}{A vector of methylated site locations (in base pairs). Elements
should be of the same order as the columns of M.}

\item{nbasis}{The number of B-spline basis functions (default is cubic
splines). Note that nbasis should not be smaller than norder (default is 4).}

\item{start}{Start location of the region.}

\item{end}{End location of the region.}

\item{...}{Optional arguments for \code{create.bspline.basis}.}
}
\value{
A matrix of basis function values. A matrix with dimensions n by Kb,
where Kb is the number of basis functions.
}
\description{
Methylation Levels to Basis Function Values
}
\examples{
M <- ar_dense$methyl
pos <- ar_dense$pos
l <- length(pos)

## B-spline basis functions
X <- basisMat(M, pos, nbasis = 50, norder = 3)
}
\seealso{
\code{\link[fda]{create.bspline.basis}}
}
