% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{bs_dense}
\alias{bs_dense}
\title{Simulated Methylation Data with B-spline Effects in Dense Scenario}
\format{
A list object.
\describe{
  \item{trait}{A vector of length 500 containing disease labels
  (250 cases and 250 controls).}
  \item{methyl}{A 500 by 100 matrix containing methylation levels. Each
  row represents the methylation profile for one subject. Each
  element of the matrix denotes BS-seq methylation level at a CpG site.}
  \item{pos}{Locations for CpG sites (in base pairs).}
}
}
\usage{
data(bs_dense)
}
\description{
A simulated dataset containing 500 subjects and 100 methylation sites.
Methylation effects are smoothed by B-spline basis functions.
20\% of the basis functions are simulated to be causal/associated. Effect
strengths are randomly sampled from U(-0.05, 0.05) distribution.
}
\keyword{datasets}
