% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{fourier_sparse}
\alias{fourier_sparse}
\title{Simulated Methylation Data with Fourier Effects in Sparse Scenario}
\format{A list object.
\describe{
  \item{trait}{A vector of length 1,000 containing disease labels
  (523 cases and 477 controls).}
  \item{methyl}{A 1,000 by 490 matrix containing methylation levels. Each
  row represents the methylaiton profile for one subject. Each
  element of the matrix denotes BS-seq methylation level at a CpG site.}
  \item{pos}{Locations for CpG sites (in base pairs).}
}}
\usage{
data(fourier_sparse)
}
\description{
A simulated dataset containing 1,000 subjects and 490 methylation sites.
Methylation effects are smoothed by Fourier basis functions.
1\% of the basis functions are simulated to be causal/associated. Effect
strengths are randomly sampled from U(-5, 5) distribution.
}
\keyword{datasets}
